vision_experiments/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bidict==0.23.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
eventlet==0.33.3
filelock==3.13.1
Flask==1.1.1
Flask-SocketIO==4.1.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.2.0
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==3.13.0
python-socketio==4.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.13.1
simple-websocket==1.0.0
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.14.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==2.0.3
wrapt==1.14.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
